+++ src/AutoloadDevSection.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Nadar\PhpComposerReader;
+
+
+/**
+ * Require Section Iterator.
+ *
+ * @author Basil Suter <basil@nadar.io>
+ * @since 1.3.0
+ */
+class AutoloadDevSection extends AutoloadSection
+{
+    const SECTION_KEY = 'autoload-dev';
+}