+++ tests/AutoloadSectionTest.php
@@ -5,6 +5,7 @@ namespace Nadar\PhpComposerReader\Tests;
 use Nadar\PhpComposerReader\ComposerReader;
 use Nadar\PhpComposerReader\AutoloadSection;
 use Nadar\PhpComposerReader\Autoload;
+use Nadar\PhpComposerReader\AutoloadDevSection;
 
 class AutoloadSectionTest extends ComposerReaderTestCase
 {
@@ -18,6 +19,19 @@ class AutoloadSectionTest extends ComposerReaderTestCase
             $this->assertSame('Nadar\\PhpComposerReader\\', $autoload->namespace);
             $this->assertSame('src/', $autoload->source);
             $this->assertSame(AutoloadSection::TYPE_PSR4, $autoload->type);
+        }
+    }
+
+    public function testDevIterator()
+    {
+        $reader = new ComposerReader($this->getValidJson());
+        $section = new AutoloadDevSection($reader);
+        
+        foreach ($section as $namespace => $autoload) {
+            $this->assertSame('TEST', $namespace);
+            $this->assertSame('TEST', $autoload->namespace);
+            $this->assertSame('test/', $autoload->source);
+            $this->assertSame(AutoloadDevSection::TYPE_PSR4, $autoload->type);
         }
     }